requirements.txt
torch
torchvision
pynput
mss
opencv-python
Pillow
numpy
